packages/mocks/pyproject.toml
[tool.coverage.run]
branch = true
source = ["mocks"]

[tool.coverage.report]
exclude_lines = ["if TYPE_CHECKING:"]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/mocks/html' --cov-report xml:'../../coverage/mocks/coverage.xml' --html='../../reports/mocks/unittests/html/index.html' --junitxml='../../reports/mocks/unittests/junit.xml'"

[tool.poetry]
name = "mocks"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = []
license = "Proprietary"
readme = "README.md"

[[tool.poetry.packages]]
include = "mocks"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
transformers = "^4.49.0"
torch = [
    { version = "^2.6.0+cpu", source = "pytorch_cpu", platform = "linux", markers = "sys_platform != 'darwin'" },
    { version = "^2.6.0+cpu", source = "pytorch_cpu", platform = "win32", markers = "sys_platform != 'darwin'" },
    { version = "^2.6.0", source = "pypi", platform = "darwin", markers = "sys_platform == 'darwin'" },
]
accelerator = "^2024.9.13"
fastapi = "^0.115.11"
google-generativeai = "^0.8.4"
uvicorn = "^0.34.0"

[tool.poetry.dependencies.common]
path = "../common"
develop = true

[tool.poetry.group.dev.dependencies]
transformers = "^4.49.0"
folktables = "^0.0.12"
scikit-learn = "^1.6.1"
torch = [
    { version = "^2.6.0+cpu", source = "pytorch_cpu", platform = "linux", markers = "sys_platform != 'darwin'" },
    { version = "^2.6.0+cpu", source = "pytorch_cpu", platform = "win32", markers = "sys_platform != 'darwin'" },
    { version = "^2.6.0", source = "pypi", platform = "darwin", markers = "sys_platform == 'darwin'" },
]

[tool.poetry.group.train.dependencies]
torch = [
    { version = "^2.6.0+cpu", source = "pytorch_cpu", platform = "linux", markers = "sys_platform != 'darwin'" },
    { version = "^2.6.0+cpu", source = "pytorch_cpu", platform = "win32", markers = "sys_platform != 'darwin'" },
    { version = "^2.6.0", source = "pypi", platform = "darwin", markers = "sys_platform == 'darwin'" },
]
torchvision = [
    { version = "^0.21.0+cpu", source = "pytorch_cpu", platform = "linux", markers = "sys_platform != 'darwin'" },
    { version = "^0.21.0+cpu", source = "pytorch_cpu", platform = "win32", markers = "sys_platform != 'darwin'" },
    { version = "^0.21.0", source = "pypi", platform = "darwin", markers = "sys_platform == 'darwin'" },
]
torchaudio = [
    { version = "^2.5.2+cpu", source = "pytorch_cpu", platform = "linux", markers = "sys_platform != 'darwin'" },
    { version = "^2.5.2+cpu", source = "pytorch_cpu", platform = "win32", markers = "sys_platform != 'darwin'" },
    { version = "^2.5.2", source = "pypi", platform = "darwin", markers = "sys_platform == 'darwin'" },
]


[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
